requirements.txt
fastapi==0.104.1
mangum==0.17.0
transformers==4.36.0
python-dotenv==1.0.1
boto3==1.24.53
numpy>=1.24.3,<2.0
insightface==0.7.3
python-multipart==0.0.6
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.1.0+cpu
torchvision==0.16.0+cpu
opencv-python-headless==4.8.1.78
onnxruntime==1.16.3
onnx==1.15.0
pillow==10.0.1
tqdm==4.66.1 
elasticsearch==8.10.0